adds 'Visit DAO' button to details - extracts AddressDisplay components to own file
